refactor into new hooks
